session_09/pyproject.toml
[project]
name = "session_09"
version = "0.1.0"
description = "A simple Discord bot"
authors = [
    {name = "cstriker421",email = "cstriker421@gmail.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
discord = "^1.7.3"
typer = "^0.4.0"
python-dotenv = "^0.19.2"
python = ">=3.9"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2"
pytest-asyncio = "^0.15"

[tool.poetry]
packages = [
    { include = "session_09", from = "src" }
]